Fold RelationshipInterpreter into RelationshipTracker.
The extra inheritance just added extra complexity without
benefit and the child class actually referenced the parent
in one of the methods.

Note: This is a save-breaking changei. We can avoid that by
still keeping the parent class around but making it an empty
class.
